+++ src/app/records/presentation-nodes.ts
@@ -1,11 +1,11 @@
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
+import { t } from 'app/i18next-t';
 import { DimItem } from 'app/inventory/item-types';
 import { ItemCreationContext, makeFakeItem } from 'app/inventory/store/d2-item-factory';
 import { ItemFilter } from 'app/search/filter-types';
 import { compact, count, filterMap } from 'app/utils/collections';
 import extraItemCollectibles from 'data/d2/unreferenced-collections-items.json';
 
-import { t } from 'app/i18next-t';
 import { DimTitle } from 'app/inventory/store-types';
 import { getTitleInfo } from 'app/inventory/store/d2-store-factory';
 import { compareBy } from 'app/utils/comparators';
@@ -509,9 +509,31 @@ export function toRecord(
   }
 
   // Rename Immovable Refit -> Vexcalibur Catalyst
-  if (recordHash === 3787307395) {
-    // @ts-expect-error name is a read-only property
-    recordDef.displayProperties.name = t('Records.VexcaliburCatalyst');
+  const VEXCALIBUR_CATALYST_RECORD_HASH = 3787307395;
+  if (recordHash === VEXCALIBUR_CATALYST_RECORD_HASH) {
+    Object.assign(recordDef, {
+      displayProperties: {
+        ...recordDef.displayProperties,
+        name: defs.Record.get(recordHash).stateInfo.obscuredName,
+      },
+    });
+  }
+
+  const CATALYST_PRESENTATION_NODES = [3788273704, 185103480, 2538646043];
+  if (
+    recordDef.parentNodeHashes?.length > 0 &&
+    recordDef.parentNodeHashes.some((hash) => CATALYST_PRESENTATION_NODES.includes(hash)) &&
+    recordDef.stateInfo?.obscuredDescription
+  ) {
+    const sourceText = `\n\n${t('Progress.CatalystSource', { source: recordDef.stateInfo.obscuredDescription })}`;
+    if (!recordDef.displayProperties.description.includes(sourceText)) {
+      Object.assign(recordDef, {
+        displayProperties: {
+          ...recordDef.displayProperties,
+          description: `${recordDef.displayProperties.description}${sourceText}`,
+        },
+      });
+    }
   }
 
   const trackedInGame = profileResponse?.profileRecords?.data?.trackedRecordHash === recordHash;